Clear spatial index when updating mesh coordinates (#89)
Call clear_spatial_index() on the moved mesh when updating its
coordinates in MA mover. Otherwise if a point was queried (e.g. with
Function.at()) before the update, the same spatial index will still be
used when the query is repeated after the update but at that time the
point may be located in a completely different cell resulting in
PointNotInDomainErrors.