Better testing for Monge-Ampere boundary conditions (#95)
Closes #79.
Partially addresses #19.

This PR achieves 100% code coverage for the `monge_ampere` module. In
particular, it introduces unit tests for the boundary condition logic
for the equidistribution equation. The boundary condition code is
refactored, including pulling it out as a separate method so it can be
more easily targeted.

The PR also introduces a new `math` module including functions for
generating equations of lines and planes. These are used to verify that
boundary conditions are only applied on straight line boundary segments
(2D) and planar boundary segments (3D). The new module has 100% code
coverage.

Finally, an additional error check test is added for the `tangling`
module to get that to 100% code coverage, too.