Fix PETSc errors in CI (#152)
Closes #151.

This PR addresses two issues that cropped up with recent Firedrake
updates:
* Assembling a 1-form to a tensor now requires the tensor to be
represented as a `Cofunction`.
* The Quasi-Newton solver no longer works in 1D. Looking more closely,
I'm surprised it ever did. It uses a vector space basis that accesses
`sub(1)`, which doesn't exist in 1D. I think it's best to just drop
support for the QN approach in 1D, as done here.